refactor: component 분리
#6
